feat: functional options
